// Api.ts
export interface PetDto {
  id?: string;
  name?: string;
  breed?: string;
  imgurl?: string;
  sold?: boolean;
}

export interface CreatePetDto {
  name?: string;
  breed?: string;
  imgurl?: string;
}